break out the share dialog into it's own file
